Add a patch to make BIND's work on pre-2.6.37
This commit makes a precompiled version of BIND9 and its supplementary
tools not to abort on Linux kernels older than 2.6.37.  Before that,
BIND9 and its tools could abort when establishing a TCP-based DNS
transport connection.

The problem was related to the absence of TCP_USER_TIMEOUT socket
option on older versions of Linux. The fix makes the code apply this
option on the best-effort basis rather than require it.